Convert portal licensing store to TS
