+++ packages/types/src/api/web/global/self.ts
@@ -1,4 +1,5 @@
-import { DevInfo, User } from "../../../documents"
+import { License } from "../../../sdk"
+import { Account, DevInfo, User } from "../../../documents"
 
 export interface GenerateAPIKeyRequest {
   userId?: string
@@ -9,4 +10,9 @@ export interface FetchAPIKeyResponse extends DevInfo {}
 
 export interface GetGlobalSelfResponse extends User {
   flags?: Record<string, any>
+  account?: Account
+  license: License
+  budibaseAccess: boolean
+  accountPortalAccess: boolean
+  csrfToken: boolean
 }
